+++ src/dreamcycle/sdk/__init__.py
@@ -1,12 +1,24 @@
 """Public vendor SDK."""
 
 from dreamcycle.sdk.client import DreamCycleClient, DreamCycleSDKError
-from dreamcycle.sdk.models import AdapterState, CycleJob, MemoryItem
+from dreamcycle.sdk.models import (
+    AdapterState,
+    CycleJob,
+    KnowledgeEdgeItem,
+    KnowledgeItem,
+    KnowledgeNeighbor,
+    KnowledgeStats,
+    MemoryItem,
+)
 
 __all__ = [
     "AdapterState",
     "CycleJob",
     "DreamCycleClient",
     "DreamCycleSDKError",
+    "KnowledgeEdgeItem",
+    "KnowledgeItem",
+    "KnowledgeNeighbor",
+    "KnowledgeStats",
     "MemoryItem",
 ]